Additional fixes and cleanup
