+++ docs/css/generator.css
@@ -1,0 +1,99 @@
+#generator {
+  label {
+    font-size: 24px;
+    font-weight: 400;
+    margin-bottom: 0.5em;
+    display: block;
+  }
+
+  fieldset label {
+    font-size: 20px;
+  }
+
+  fieldset {
+    border: none;
+    padding: 0;
+  }
+
+  legend {
+    font-size: 24px;
+    font-weight: 400;
+  }
+
+  .checkboxes {
+    margin-top: 8px;
+    padding-left: 24px;
+    label {
+      cursor: pointer;
+    }
+    > div {
+      margin-top: 2rem;
+      margin-bottom: 1rem;
+    }
+  }
+
+  .mb-4 {
+    margin-bottom: 24px;
+  }
+
+  .my-2 {
+    margin-top: 1rem;
+    margin-bottom: 1rem;
+  }
+
+  .text-sm {
+    font-size: 80%;
+  }
+
+  .text-red {
+    color: #eb5757;
+  }
+
+  button {
+    text-transform: none;
+    -webkit-appearance: button;
+    font-family: inherit;
+    font-size: 24px;
+    line-height: 1.15;
+    margin: 0 0 3rem 0;
+    border-radius: 1em;
+    border: none;
+    background-color: #4ba0a5;
+    padding: 1em 24px;
+    color: white;
+    cursor: pointer;
+    &:hover {
+      background-color: #00b6af;
+    }
+  }
+
+  input[type="text"],
+  input[type="url"],
+  input[type="email"],
+  select,
+  textarea {
+    font-size: 1.8rem;
+    width: 100%;
+    border-radius: 10px;
+    padding: 1rem 24px;
+    border: 2px solid #c8cdd0;
+    &:focus {
+      border-color: #00b6af;
+      outline: none;
+    }
+    &:read-only {
+      &:focus {
+        border-color: #c8cdd0;
+      }
+      font-size: 1.2rem;
+      padding: 1rem 1rem;
+    }
+  }
+
+  dt {
+    font-weight: 600;
+  }
+  dd {
+    margin-bottom: 3rem;
+  }
+}
